clean code
add OLD calculation. Dont merge this, used now to test speed differences, old method seems faster with the test dataset.
